document multi-currency and regenerate sdks with currency fields

+++ others/python-sdk/src/autumn_sdk/models/trackop.py
@@ -91,7 +91,7 @@ class TrackParamsTypedDict(TypedDict):
     r"""The amount of usage to record. Defaults to 1. Use negative values to credit balance (e.g., when removing a seat)."""
     properties: NotRequired[Dict[str, Any]]
     r"""Additional properties to attach to this usage event."""
-    timestamp: NotRequired[float]
+    timestamp: NotRequired[int]
     r"""Unix timestamp in milliseconds to use for the usage event. Defaults to the current time."""
     async_: NotRequired[bool]
     r"""If true, enqueue the event for asynchronous processing and return 204 immediately. The response will not include balance information."""
@@ -117,7 +117,7 @@ class TrackParams(BaseModel):
     properties: Optional[Dict[str, Any]] = None
     r"""Additional properties to attach to this usage event."""
 
-    timestamp: Optional[float] = None
+    timestamp: Optional[int] = None
     r"""Unix timestamp in milliseconds to use for the usage event. Defaults to the current time."""
 
     async_: Annotated[Optional[bool], pydantic.Field(alias="async")] = None
